Update quoted-printable encoding
Research has been extensively done and Blizzard does things weird, but
usefully weird. ALL valid UTF-8 characters consisting of bytes at or
above 0x80 are permitted, even if they're not printable (due to absence
of glyph, due to being undefined, or due to being control characters).
ASCII control characters are disallowed, except 0x09 and 0x7f (for some
reason).

So just encode invalid UTF-8 byte sequences, and the disallowed control
characters. And make sure to not split a message along either a
quoted-printable character (for on-the-fly reconstruction on the other
end) or along a UTF-8 sequence (or the message silently fails).

This is probably buggy and definitely needs a lot more testing.