Rename AddOn_Chomp.GetCapability() to CheckReport()
This function will replace the capability checker and serve as a
preliminary function to actually sending a report, by returning two
values: first, whether a report can be made (always false for BattleNet
and InGame/unlogged messages), and then the message method preferred for
the target. It's missing the check against Blizzard's own function
currently, but that will be added later.