Add preliminary reporting API
This API (undocumented so far) permits checking whether a subject can be
reported for various reasons (known/unknown GUID, BattleNet or Unlogged
message types, and if Blizzard's API says they can be reported), and
then permits the filing of an automatic report with a brief reference to
the objectionable material being in the addon data channels.

Future plans are to expand this to allow custom report reasons and more
detail.