Remove forceMethod, rearrange smart message function to account
forceMethod could create undesirable situations (messages being dropped when it would be possible to deliver normally), so ditch it to avoid confusion.

Additonally, return only a string with the method used to send the message, rather than three booleans.

And finally, use the standard channel queue name for universalBroadcast message.