Compatibility change for the future
In the future (months to years), the use of 0x7f will be stopped, in case Blizzard yanks the rug out from under a non-printable character in logged addon messages, and replaced by the colon. This change enables that change to be made in the future without breaking any clients using this version of Chomp or later.