Make serialization a bit safer
Serialization could previously return a function inside a table (with an unusuably barren env, but still). Now, if the word 'function' is found in the string to deserialize, the entire table is checked recursively for any forbidden types.

Exploiting this, in any way, would have required finding a way to exploit the empty environment and required the embedding addon to do some really bad things too, but it's good to fix stuff like this anyway.