Improvements to serialization efficiency
The first batch of changes is modifications to how tables are
serialized. Previously, we'd always quote string keys in tables
even if the key is a valid Lua identifier. We now check if the
key can be represented without brackets and, if so, omit them.
This saves ~4 bytes per string table key that's eligible, and
costs nothing to ones that aren't. A lot of the data we send
fits this quite nicely.

The next change is to serialize tables in two passes; the first
does the array portion (from t[1] through t[n-1], where n is the
index of the first nil value. These are written without keys. The
second pass does the rest of the table, skipping integral keys within
the already-written range.

These changes don't require any modification to the logic for
deserialization as the resulting string is still a valid Lua
value for loading. Alone they are backwards compatible.

The final change which is *not* backwards compatible is to use the
grave (`) as our escape character for quoted-printables instead of
"=", since "=" is an extremely common character for serialized data
and is also prevalent in the MSP protocol. The "`" character should
be rare enough in profiles and isn't a valid token in Lua, so won't
expand out to 3 bytes as often as "=" would. This also has the
benefit of making the encoded data a lot more readable for logged
comms.