Migrate to 8.2.5 C_BattleNet APIs for 9.x
This replaces a couple of usages of deprecated APIs with the ones
added in 8.2.5. For older clients, the legacy API is used but its
return values are wrapped into the structure provided by the newer
API.

Additionally to work around memory churn a basic cache of data
has been added for account info returns; this cache is eagerly
pruned upon so much as a slight breeze to the BN friends API
data, so shouldn't cause any issues.