Allow BNet to match accounts case-insensitively
This improves the internal BNGetIDGameAccount function to compare
the name/realms of candidate accounts in a case insensitve manner
to what the caller has requested; so if we want to send a message
to "someone-kirintor" then we'll correctly find their account in
the Battle.net friends list.

A couple of utility string functions are made public for this, one
is a case-insensitive string comparison function that will prefer
a basic equality test first, followed by using either the UTF8 lib
if present, and then falling back to the (locale-dependant, likely
ASCII-only) string library.

Additionally a function to split a merged name/realm is added as a
utility bonus, since we need it in TRP and rewriting it a million
times isn't fun :)