Change escape character to "~"
The previous choice of "`" made for 9.0 was a bit rushed and was known
to conflict with LibMSP's separator. The idea we had at the time was to
change LibMSP to not be quite so reliant upon the separator, but doing
so has other issues which would require larger changes to its own
protocol.

As a compromise instead, this changeset swaps the use of "`" to a "~"
which doesn't conflict with either TRP or MSP comms and should prove
to be optimal.