Implement a soft-upgrade for the encoding change
This takes advantage of a bug (or a gift!) in existing versions of Chomp
where the 8th unused bit in comms was typoed to "UNUSES8". This gives us
a way to quietly signal support for a protocol upgrade without outright
breaking comms for a user.

The encoding character change now coexists with the old one and has been
split into a version 1 and version 2 codec. By default, all
communications use codec version 1 and thus the old "`" character.

When a user with Chomp v16+ sends a message, we set bit 8 which has been
renamed to "VERSION16" and will be quietly ignored by older clients on
receipt. When this is received we cache information on a per-sender
level indicating that the player who sent the message is running a new
version.

When sending data via SmartAddonMessage to a target that has previously
indicated v16+ support, we upgrade the codec to version 2 and set a
separate bit indicating that it is now in use.

This means that we've got a way to have the two escape sequences coexist
side by side, with the older less inefficient one being the default but
comms upgrading to the newer one if found to be supported.

There is one edge case of a user who initially logs in with v16+ of
Chomp, sends messages, and then reloads with an older version - in that
case they'd get a warning and drop the incoming messages from anyone
who tries to use codec version 2. We do however clear the support flag
in such a case, so comms won't break _forever_ for them - they would
just need to send a message back to the original sender without the
VERSION16 bit set.