Add initial packager support
This implements support for the packager and uploading new zips to
GitHub releases, as well as packaging internal dependencies which
is going to be needed for a few upcoming changes to the library.

This is the _minimal_ work for the packager as a standalone reviewable
unit.

Of note, Chomp now embeds ChatThrottleLib so that it can ensure 100%
of the time is loaded for its own redirection layer. I'm still hoping
that in the future we can just kill that part and have CTL do all the
backend comms.