Remove API documentation integrations
The ingame API documentation system is seldom actually used for anything
outside of Blizzard and we've not been keeping this documentation up
to date as-is, in addition to it being wrong in a few areas at this
point as well.

As it's not really adding any value this commit just yeets it. We can
possibly revisit shoving it in the README at a later date, but until
then no one is really using the library enough to justify maintaining
the documentation, and most authors are already perfectly acquainted
with the idea of just reading the implementation - which is always
accurate, up-to-date, and has self-documenting errors and assertions.