Add generic callback registry support
This adds a callback registry (provided by CallbackHandler) to provide
a way to implement a few features that have been desirable for a small
while now.

The callback API exposed by the library is a sane combination of the
one offered by Blizzards' CallbackRegistryMixin and CallbackHandler:

-  AddOn_Chomp.RegisterCallback(event, func, owner)
-  AddOn_Chomp.UnregisterCallback(event, owner)
-  AddOn_Chomp.UnregisterAllCallbacks(owner)

RegisterCallback will register the callback (func) for the named event
and bind it to the given owner. The owner can be a string, table, or
thread and is synonymous with the "self" used by CallbackHandler.

The callback will be dispatched with the owner as the first parameter,
followed by all arguments supplied with the event. The event name isn't
passed.

UnregisterCallback and UnregisterAllCallbacks act the same as their
counterparts in CallbackHandler, unregistering either the single
handler associated with an event for a given owner, or all events for
a given owner.

With this in place, the events that have been defined are as follows:

- OnError
- OnMessageReceived

The OnError event replaces the existing ErrorCallback mechanism and
acts as it does today. The existing error callback related APIs
internally register to this event now, so backward compatibility
is maintained.

The OnMessageReceived event is dispatched each time any callback for
a registered prefix would be executed, with the same arguments supplied
to the prefix callback. This solves a longstanding annoyance of not
being able to generically register multiple callbacks for a single
prefix, which was resolved in another branch but not merged in due to
other concerns.

Future callbacks might involve an OnMessageSent or notifications for
comms load, but we'll cross those bridges when we get there.