Use LibStub for versioning
This makes Chomp a Good Citizen (tm) and stops it from reinventing the
wheel of library versioning for no heckin' good reason.

Chomp now registers itself via LibStub with the major "Chomp" and the
same incrementing (minor) version number it always has. The internal
table is stored on our library instance as the "Internal" subkey
instead of as the "__chomp_internal" global, though for compatibility
with addons that might pull in older versions of the library the old
internal global is still exported until the PLAYER_LOGIN event has
fired.

This change is fully backwards compatible as far as the public API is
concerned; The AddOn_Chomp global is still defined and just points to
the same library table registered with LibStub. For addons that have
captured the old AddOn_Chomp as an upvalue this _shouldn't_ cause an
issue as there's no state on the old table that's required by the
library, since the whole public API avoids use of "self".

Note that the internal table is accessible after PLAYER_LOGIN has
fired via "Internal" key of the API table; this is intentional to assist
with debugging.

No work has been done to make the library upgradable after PLAYER_LOGIN
has fired; this would be a substantial task and likely can't be done
until the older versions are no longer in circulation.