Remove the "()" from all error strings
Mostly for consistency with error styles elsewhere such as in Blizzard's
newer code; the "()" for functions doesn't really add any clarity, so
might as well clean these up while we're here.

The error messages in the CTL compat section are unmodified
intentionally as these should be the same as the original lib.