Use doubly linked lists for queues and priorities
This is mostly a performance enhancement when dealing with message
queues for players that get fairly large.

The existing queue system uses an array from which queues and messages
are removed from the start which causes all remaining elements to be
down-shuffled. In the case of large data transfers send to lots of
people, this would stack up and could have a noticable impact on CPU
usage.

As the primary operation performed on our queue is either a pop-front
or push-back and random access is never needed, it makes more sense to
instead use a doubly linked list.

Mercifully Blizzard already has a mixin for this that implements an
intrusive form where the front/back/prev/next keys are all stored on
the inserted nodes directly; so this changeset just swaps our queue
implementation to using that behind the scenes.