Remove support for older serialization codec
This removes support for the older serialization codec which
was replaced a while ago due to efficency concerns over its choice
of escape character.

The support for the newer codec was added in v16 - released 12 months
ago. Compatibility at the time with pre-v16 peers was maintained through
abusing a typo in the addon that allowed us to set a 'VERSION16' bit
on messages that was ignored by older clients. If a v16+ client then
received this bit, it would mark the peer as supporting the newer codec
and use it for future transmissions indicated through a new 'CODECV2'
bit.

By removing support for the older codec, the following changes are
made:

  1. The 'CODECV2' bit has been renamed 'VERSION20'.
  2. All outgoing smart messages will set the 'VERSION20' bit.

The renaming of the bit has no effect on older clients; v16-v19 will
still see it as inferring 'CODECV2' and use the more efficient
serialization codec as a result.

The bit being set by default, however, means that there is no longer
the initial negotiation phase whereby Chomp needs to know if a peer
supports the newer codec - it will just assume that all peers on the
network are new enough.

As a result, compatibility between v16-v19 and v20 peers remains as-is.
Compatibility between pre-v16 and v20 peers is removed, which will
result in the following:

  - If v20+ receives a message from a pre-v16 peer, the message will be
    silently discarded.

  - If pre-v16 peer receives a message from a v20+ peer, the 'VERSION20'
    bit will be set which triggers the "unused bit" logic in the older
	versions leading to the message being discarded and the user told
	that their version of Chomp is too old to support whatever they
	just received.