Use v2 serialization codec by default
This removes the ability for us to communicate with pre-v16 versions
of Chomp by enforcing that all serialization goes through the new
codec version. This will, in turn, allow us to deprecate and remove
the old v1 codec later on down the line when nothing on the network
should be sending us v1 codec data. Unfortunately that can't be done
now due to the nature of codec autonegotiation starting off with
v1 and upgrading to v2 based upon received bitflags.