Remove embedded libraries from XML file
Embedded libraries are now only loaded for packaged builds, and
consumers of the library need to embed our dependencies and make
them available ahead of loading Chomp.

This follows the general practices of other libraries and solves the
needless duplication of multiple copies of the same embedded libraries
in a single addon.