Route all messages through ChatThrottleLib
This removes all internal management of internal transmission queues,
leaving Chomp as a wrapper that really just provides "smart" routing with
the worst implementation of message serialization known to mankind.

The upstream CTL library has at this point merged in all the support
required for sending logged messages. Battle.net message support is
still pending but queued up.

One thing that we still provide for now is an increase to CTL's ancient,
outdated, and totally crappy default transmission rates. Hopefully we
can just get newer happy values upstreamed at some point, but that might
be a bit of an ask in the face of all the other major changes made to
appease Blizzard's new per-prefix throttles.