Improve handling of realm names in message filter
When a WHISPER message is sent to a player with a fully qualified name
that includes the realm, the error message sent back by the server won't
necessarily be 1:1 with the string submitted as the "target" parameter
of the API.

There's two cases where this happens:

  1. The realm name wasn't in normalized form ("Azjol Nerub")
  2. The realm name wasn't case-correct ("azjolnerub")

In both of the above cases the error message given by the server would
case-correct and normalize the realm name in the response, leading to
cases where the key in our target table doesn't align with that of the
fully qualified player name received in the error message, resulting
in the offline player message filter potentially allowing errors
through.

This changeset attempts to paper over the issue by normalizing target
names client-side; we'll transform the target name to a lowercase format
and normalize the realm segment if present and store that as the key
for our filter. The same key is then generated and looked up on receipt
of an error, which should result in more correct handling if the server
applied any such transforms itself.