Only use character names for error filtering
Chomp (and a lot of other RP-related stuff) incorrectly use the
localized realm name in a lot of stuff relating to comms rather
than normalized realm names.

For realms in the latin alphabet this works fine largely as we've been
performing rudimentary normalization, however for realms in non-latin
alphabets this approach for normalization doesn't work.

As changing to normalized realm names would be a wide-ranging and
broadly painful change, the message filter for offline player messages
will now only use the lowercased character name as the filter key.

This should be fine as it's not likely that someone will be sending
messages to two identically named characters on different realms at
once.