Handle throttling in immediate-send logic
There's a few naked calls to SendAddonMessage APIs that occur if there's
no queued data. These could potentially return throttle error codes, so
we need to handle those.

If these calls indicate that the data the user wants to send is
throttled, we won't return early and will instead trigger the queue
logic.

Note also that this code always seems to invoke the callback with a
true 'didSend' parameter and never bothered to check for a false return
previously. Won't rock the boat and fix this now though.