Avoid potential select errors
The despooling logic is generic over which send function it executes -
it simply calls a queued 'f' and unpacks a bunch of parameters into it.

Not all 'f's are the same though when it comes to return values, and it
may be the case that one f returns a boolean, another f returns a result
code, and another f returns nothing whatsoever.

In the nothing whatsoever case, select will explode due to there being
no results on the stack - so we'll handle that by assuming that any f
that returns absolutely nothing sent its data successfully.