Remove support for older serialization codec (#47)
This removes support for the older serialization codec which
was replaced a while ago due to efficency concerns over its choice
of escape character.

The support for the newer codec was added in v16 - released about three
years ago. Compatibility at the time with pre-v16 peers was maintained
through abusing a typo in the addon that allowed us to set a 'VERSION16'
bit on messages that was ignored by older clients. If a v16+ client then
received this bit, it would mark the peer as supporting the newer codec
and use it for future transmissions indicated through a new 'CODECV2'
bit.

About two years ago, we made a change to drop pre-v16 support by
unilaterally broadcasting the 'CODECV2' bit, and to ignore all messages
from ancient clients.

Now we're safely in a position where the old codec can be entirely
removed. This commit makes no changes to the wire protocol, nor does it
alter any bits - it strictly removes a rather large dead code path that
is entirely unreachable.