Add support for sending messages to battletags
This is half the support required to get Chomp capable of sending to
accounts matched by a battletag, which could be used to enable comms
between client flavors on an opt-in basis.

There's work that needs solving on the receiver half unfortunately, as
we need to synthesize a player name of some kind.