Small tweaks to bandwidth constants
Nerfing our burst figure and message overhead adjustments slightly.

The burst figure this has been lowered to is one that I'm proposing for
upstream CTL usage, and is the midpoint between its own current figure
and ours. Burst is the one that's usually the most likely to make the
server angry and disconnect you, so having this be a bit lower makes
sense from a safety perspective.

The CPS value remains the same, as this is the one that we honestly
really care about - however this may get lowered to 1536 in due course.

For message overhead, our 27 byte figure isn't sane. The client packet
used for comms messages includes in the worst case a 128-bit UUID for a
channel identifier, plus about 8 bytes of additional overhead for the
rest of the packet _and_ we also need to factor in the target name and
prefix.

Really the overhead should be higher but it amortizes out over time -
but either way, 27 bytes assumes most messages are broadcasts rather
than whispers - so let's bump it just a tiny bit.