Remove broadcast 'upgrade' support (#58)
Chomp's "smart" routing feature allows a prefix to opt-in to having its
messages yeeted out across group chats if the destination target for a
whisper is in your group. This is strictly an optimistic upgrade and not
a guarantee on comms routing.

With the advent of per-prefix throttles and the difference between
behavior for whispers vs. group channels, this ends up being a trap that
can ultimately slow such comms down to a crawl.

For now let's kill off this support from the perspective of senders
entirely and revisit if it's desirable later. Chomp can still receive
messages with the broadcast bit set and process them accordingly, but
going forward it won't send them out.