Cap target length to 47 bytes
This works around an instant disconnect bug with SendAddonMessage when
sending whispers to players whose combined "character-realm" name
appears to exceed this threshold, most easily noticed with players
from Russian realms.

As the issue is being actively fixed, hopefully we don't need this
workaround for too long - so a time bomb is included to disable it
after about two weeks in the event a user doesn't update.

This is only applied to "smart" messages, as those are the only
ones guaranteed to be "name-realm" qualified.