Add C_EncodingUtil support
This extends Chomp's "smart" addon messaging with support for
alternative serialization methods (CBOR and JSON) using Blizzard's new
APIs where available, as well as transparent compression and binary
encoding of data.

For serialization, the behavior is such that if the message options use
today's 'serialize' boolean then the serialization behavior you'll get
will be a suitable default chosen by Chomp. This is unconditionally the
(existing) Lua data serializer, but long-term will be changed to CBOR
when this new version is widely available.

Specific serialization methods can be selected by setting the
'serialize' option to a string value; any of "CBOR", "JSON" or "LUA" are
accepted.

In the case of CBOR or JSON encoding, Chomp will detect if the client is
capable of supporting the requested option and will _silently_ fall back
to Lua if not available. This makes it possible for addons to just say
that they'd prefer a more optimal encoding if available without having
to handle fallback logic themselves.

Compression is a new addition and behaves similarly; if the 'compress'
option on a message is set to 'true' then you'll get an appropriate
default method chosen by the library. Otherwise, you can force selection
of a compression algorithm - though currently the only valid option here
is "DEFLATE".

As both CBOR serialization and compression produce binary output, if
either of those are requested for a message then Chomp will
automatically perform binary encoding on the data to escape embedded
null bytes, and will reverse that upon receipt of the message.