Don't check for CODECV2 version bits on receipt
These bits have been unilaterally set on all messages since Chomp v20+.

It'd be nice to reclaim these bits for the message header, so we should
stop checking that they're set and assume we live in a modern world.
Longer term (~12.0) we can stop setting them entirely, and a bit after
that we can rename the bits and mark them unused again.